test: add multi-user integration tests
1. Test that multiple users can be registered without clobbering each
   other
2. Test that one user can change their password without it affecting the
   other

I'd like to also test sync with multiple users, to ensure we never
accidentally leak data cross-users.
